pyproject.toml
[tool.poetry]
name = "la-france-poems"
version = "0.1.0"
description = "Pipeline to perform sentiment analysis of my La France poem collection"
authors = ["Yifei Chen <yifeichen.github@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.120.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
langdetect = "^1.0.9"
python-dotenv = "^1.1.0"
dateparser = "^1.2.1"
google-cloud-storage = "^3.1.0"
transformers = "^4.52.3"
torch = "^2.7.0"
accelerate = "^1.7.0"
sentence-transformers = "^4.1.0"
streamlit = "^1.45.1"
matplotlib = "^3.10.3"
umap = "^0.1.1"
umap-learn = "^0.5.7"
requests = "^2.32.3"
openai = "^1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
